Bug Fixes, Reorganization
I fixed a bunch of issues with the tasks and timer so now everything
runs a lot more smoothly.  There is still a visual bug with the timer
but I am working on a new timer to fix that.  I also split the code up
into many smaller files instead of one huge one.